+++ src/lib.rs
@@ -23,7 +23,6 @@
 #![allow(non_camel_case_types)]
 
 extern crate libc;
-extern crate custom_error;
 extern crate ascii;
 extern crate semver;
 extern crate errno;