+++ src/mouse/ncurseswmouseerror.rs
@@ -1,7 +1,7 @@
 /*
     src/mouse/ncurseswmouseerror.rs
 
-    Copyright (c) 2019 Stephen Whittle  All rights reserved.
+    Copyright (c) 2019, 2020 Stephen Whittle  All rights reserved.
 
     Permission is hereby granted, free of charge, to any person obtaining a copy
     of this software and associated documentation files (the "Software"),
@@ -22,10 +22,13 @@
 
 use std::num;
 
-custom_error::custom_error! {
+use thiserror::Error;
+
 /// NCursesw mouse errors.
-#[derive(PartialEq, Eq)]
-pub NCurseswMouseError
-    LibraryError { func: String, rc: i32 } = "nmouse::{func}() (#{rc})",
-    IntError { source: num::TryFromIntError } = "{source}"
+#[derive(Error, Debug, PartialEq, Eq)]
+pub enum NCurseswMouseError {
+    #[error("nmouse::{func}() (#{rc})")]
+    LibraryError { func: String, rc: i32 },
+    #[error("{source}")]
+    IntError { #[from] source: num::TryFromIntError }
 }