+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All breaking changes are marked with [BC] and potentially require API consumer c
 
 ## [0.6.0] - ????.??.??
 - Upgraded source code to rust 2018 edition.
+- Now uses `thiserror` crate instead of `custom_error` crate to create `Error` types.
 - Changed signature of `getsyx() -> Result<Origin, NCurseswError>` to `getsyx() -> Result<Option<Origin>, NCurseswError>` to return a `None` instead of `Origin { y: -1, x: -1 }`. [BC]
 - Changed signature of `setsyx(origin: Origin) -> Result<(), NCurseswError>` to `setsyx(origin: Option<Origin>) -> Result<(), NCurseswError>`. [BC]
 - Changed signature of `intrflush()` and `intrflush_sp()` to ignore `handle/window` parameter as this is ignored in the NCurses library. [BC]